Set login_url
At some point, JupyterHub changed to take this as the POST endpoint
for the login form.  If we have an admin authenticator, it can (and
does, in the case of the GoogleOAuthenticator) change the login_url.
So, we'll set it back to the base Authenticator value, so that
username/password submissions actually get routed to the authenticate
method.

This means that when we want to redirect to the admin authenticator
login_url, we need to reference the super class specifically.